requirements.txt
asgiref==3.4.1
distlib==0.3.4
Django==3.2.13
django-filter==21.1
django-pandas==0.6.6
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
filelock==3.4.1
importlib-metadata==4.8.3
importlib-resources==5.4.0
numpy==1.19.5
pandas==1.1.5
platformdirs==2.4.0
psycopg2-binary==2.9.3
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.1.1
virtualenv==20.14.1
zipp==3.6.0
